Move functions to media interface

#20
